+++ lib_xs3_math/api/xs3_math.h
@@ -23,6 +23,8 @@ extern "C" {
 
 #include "vect/xs3_vect_s32.h"
 #include "vect/xs3_vect_s16.h"
+#include "vect/xs3_vect_s8.h"
+#include "vect/xs3_mixed.h"
 #include "vect/xs3_fft.h"
 #include "vect/xs3_filters.h"
 #include "xs3_util.h"