+++ test/vect_tests/src/main.c
@@ -12,8 +12,6 @@ int main(int argc, char** argv)
 {
     UNITY_BEGIN();
 
-
-
     CALL(test_CLS_funcs);
     CALL(test_HR_funcs);
     CALL(test_xs3_vpu_scalar_ops_s8);
@@ -45,15 +43,18 @@ int main(int argc, char** argv)
     CALL(test_xs3_sum_complex);
     CALL(test_xs3_vect_complex_s32_to_complex_s16);
     CALL(test_xs3_vect_complex_s16_to_complex_s32);
-    CALL(test_xs3_filter_fir_s32);
-    CALL(test_xs3_push_sample_s16);
-    CALL(test_xs3_filter_biquad_s32);
     CALL(test_xs3_abs_sum);
     CALL(test_xs3_max_min);
     CALL(test_xs3_energy);
     CALL(test_xs3_sqrt_vect);
     CALL(test_xs3_inverse_vect);
 
+    CALL(test_xs3_vect_s16_extract);
+
+    CALL(test_xs3_mat_mul_s8_x_s8);
+    CALL(test_xs3_mat_mul_s8_x_s16);
+
+    CALL(test_xs3_vect_s8_boolean);
 
     return UNITY_END();
 }